requirements/test.txt
pytest-playwright==0.3.3
pytest-asyncio==0.21.0
pytest-html==3.2.0
jsonschema==4.17.3